+++ package.json
@@ -40,21 +40,24 @@
     "redux": "^3.3.1"
   },
   "dependencies": {
-    "hoist-non-react-statics": "^1.0.5",
+    "hoist-non-react-statics": "^2.2.2",
     "invariant": "^2.2.0",
     "lodash": "^4.6.0",
     "preact-shallow-compare": "^1.0.1"
   },
   "devDependencies": {
     "babel-cli": "^6.6.0",
-    "babel-plugin-lodash": "^2.1.0",
+    "babel-plugin-lodash": "^3.2.11",
     "babel-plugin-transform-es2015-modules-commonjs": "^6.6.3",
     "babel-plugin-transform-object-rest-spread": "^6.5.0",
     "babel-preset-es2015": "^6.6.0",
     "babel-register": "^6.7.2",
     "expect": "^1.16.0",
-    "jsdom": "^8.1.0",
-    "mocha": "^2.4.5",
+    "jsdom": "^11.1.0",
+    "mocha": "^3.5.0",
+    "react": "^15.6.1",
+    "react-redux": "^5.0.6",
+    "redux": "^3.3.1",
     "rimraf": "^2.5.2"
   }
 }
